Updated websocket message handlers

+++ server/src/websockets-handlers/delete-message.ts
@@ -4,13 +4,38 @@ import { z } from "zod";
 import { deleteMessageSchema } from "../zod/schemas";
 
 export async function handleDeleteMessage(
-  wsUserMap: Map<WebSocket, string>,
+  ws: WebSocket,
+  userSocketMap: Map<string, WebSocket>,
   prisma: PrismaClient,
   jsonData: any
 ) {
   const messageData = deleteMessageSchema.parse(jsonData);
 
   console.log("Got delete message data from client:", messageData);
+
+  // fetch the chat to add this message to
+  const chat = await prisma.chat.findUnique({
+    where: { id: messageData.chatId },
+    select: {
+      id: true,
+      members: { select: { id: true } },
+    },
+  });
+
+  // check if this chat exists
+  if (!chat) {
+    throw new Error(
+      `VIOLATION: User ID ${ws.userId} tried deleting message from non-existent chat`
+    );
+  }
+
+  // check if this user belongs to this chat
+  const memberIds = chat.members.map((m) => m.id);
+  if (!memberIds.includes(ws.userId)) {
+    throw new Error(
+      `VIOLATION: Non-member ID ${ws.userId} tried deleting message from chat ${chat.id}`
+    );
+  }
 
   // delete message from Database
   const deletedMessage = await prisma.message.delete({
@@ -44,9 +69,13 @@ export async function handleDeleteMessage(
   };
 
   // Boardcast to all clients connected to this chat
-  const clientUserIds = deletedMessage.chat.members.map((m) => m.id);
-  wsUserMap.forEach((userId, ws) => {
-    if (!clientUserIds.includes(userId)) return;
+  const onlineClients: WebSocket[] = [];
+  deletedMessage.chat.members.forEach((m) => {
+    const client = userSocketMap.get(m.id);
+    if (client) onlineClients.push(client);
+  });
+
+  onlineClients.forEach((ws) => {
     if (ws.readyState === WebSocket.OPEN) {
       ws.send(JSON.stringify(dataToSend));
     }
